svg/painting/parsing/color-interpolation-valid.svg

<svg xmlns="http://www.w3.org/2000/svg" xmlns:h="http://www.w3.org/1999/xhtml" width="800" height="800"><title>SVG Painting: parsing color-interpolation with valid values</title><metadata><h:link href="https://svgwg.org/svg2-draft/painting.html#ColorInterpolationProperty" rel="help"/><h:meta content="color-interpolation supports the full grammar 'auto | sRGB | linearRGB'." name="assert"/></metadata><h:script src="/resources/testharness.js"/><h:script src="/resources/testharnessreport.js"/><h:script src="/css/support/parsing-testcommon.js"/><script><![CDATA[

test_valid_value("color-interpolation", "auto");
test_valid_value("color-interpolation", "srgb");
test_valid_value("color-interpolation", "linearrgb");

  ]]></script></svg>
