LayoutTests/platform/mac/imported/w3c/web-platform-tests/svg/painting/parsing/color-interpolation-valid-expected.txt


PASS e.style['color-interpolation'] = "auto" should set the property value
FAIL e.style['color-interpolation'] = "srgb" should set the property value assert_equals: serialization should be canonical expected "srgb" but got "sRGB"
FAIL e.style['color-interpolation'] = "linearrgb" should set the property value assert_equals: serialization should be canonical expected "linearrgb" but got "linearRGB"
